Added stats controller
