+++ src/http/controllers/index.ts
@@ -2,7 +2,8 @@ export * from './accounts.controller';
 export * from './beams.controller';
 export * from './blocks.controller';
 export * from './core.controller';
+export * from './governance.controller';
 export * from './health.controller';
+export * from './stats.controller';
 export * from './transactions.controller';
 export * from './validators.controller';
-export * from './governance.controller';
